re#1551 - some changes for Search widgets Layouts. Also the style sheet of search widgets combo box's  is changed, now the combo box's pop-up view's  scroll bar shows up and down arrow buttons. The data type of ICat investigation id in table workspace changed from long long to int64_t Now ICat Search works in Linux ,I need to check this works in Mac.

+++ Code/Mantid/DataObjects/src/TableColumn.cpp
@@ -13,7 +13,7 @@ DECLARE_TABLECOLUMN(double,double)
 DECLARE_TABLECOLUMN(API::Boolean,bool)
 DECLARE_TABLECOLUMN(std::string,str)
 DECLARE_TABLECOLUMN(Mantid::Geometry::V3D,V3D)
-DECLARE_TABLECOLUMN(long long,long64)
+DECLARE_TABLECOLUMN(int64_t,long64)
 
 
 } // namespace DataObjects
